Add files via upload
For this assignment you will model an object of your choosing and interest and code it in JavaScript. For example, Student, Product,  etc...You may not reuse the objects covered in class.

Requirements:
Use the constructor notation to create a “template” for the object. 25points
Create at least 2 instances of the object. 25 points
The object must have minimum of 2 properties and 1 method. 25 points
Display the properties values and method(s) result on a web page using the textContent or InnerHTML properties.  25 points